+++ include/threadschedule/detail/pool/inline_pool_backend.hpp
@@ -13,6 +13,7 @@
 
 #include "../../expected.hpp"
 #include "backend.hpp"
+#include "worker_context_guard.hpp"
 #include <functional>
 #include <future>
 #include <system_error>
@@ -148,6 +149,7 @@ public:
   void
   parallel_for_each(Iterator begin, Iterator end, F&& func)
   {
+    static_assert(detail::is_forward_iterator_v<Iterator>, "parallel_for_each requires at least a forward iterator");
     for (auto it = begin; it != end; ++it)
       func(*it);
   }